+++ ice-20201109/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-04-03 Version: 6.16.1
+- Update API DescribeAIAgentInstance: add response parameters Body.Instance.CallInfo.
+- Update API GenerateAIAgentCall: add response parameters Body.AvatarUserId.
+
+
 2026-03-31 Version: 6.16.0
 - Support API AddYikeUserCredit.
 - Support API CreateYikeProduction.